review and edit claude code
