+++ src/mcp_server_spira/features/productartifacts/tools/incidents.py
@@ -13,7 +13,7 @@ def _get_incidents_impl(spira_client, product_id: int) -> str:
 
     Args:
         spira_client: The Inflectra Spira API client instance
-        product_id: The numeric ID of the product. If the ID is PG:45, just use 45. 
+        product_id: The numeric ID of the product. If the ID is PR:45, just use 45. 
                 
     Returns:
         Formatted string containing the list of incidents
@@ -28,7 +28,7 @@ def _get_incidents_impl(spira_client, product_id: int) -> str:
 
         # Format the incidents into human readable data
         formatted_results = []
-        for incident in incidents[:25]:  # Only show first 25 incidents
+        for incident in incidents:
             incident_info = format_incident(incident)
             formatted_results.append(incident_info)
 
@@ -56,7 +56,7 @@ def register_tools(mcp) -> None:
         - Access the full description and selected fields of incidents
 
         Args:
-            product_id: The numeric ID of the product. If the ID is PG:45, just use 45. 
+            product_id: The numeric ID of the product. If the ID is PR:45, just use 45. 
         
         Returns:
             Formatted string containing comprehensive information for the
