+++ src/mcp_server_spira/features/productartifacts/tools/requirements.py
@@ -13,14 +13,14 @@ def _get_requirements_impl(spira_client, product_id: int) -> str:
 
     Args:
         spira_client: The Inflectra Spira API client instance
-        product_id: The numeric ID of the product. If the ID is PG:45, just use 45. 
+        product_id: The numeric ID of the product. If the ID is PR:45, just use 45. 
                 
     Returns:
         Formatted string containing the list of requirements
     """
     try:
         # Get the list of requirements in the product
-        requirements_url = f"projects/{product_id}/requirements"
+        requirements_url = f"projects/{product_id}/requirements?starting_row=1&number_of_rows=500"
         requirements = spira_client.make_spira_api_get_request(requirements_url)
 
         if not requirements:
@@ -28,7 +28,7 @@ def _get_requirements_impl(spira_client, product_id: int) -> str:
 
         # Format the requirements into human readable data
         formatted_results = []
-        for requirement in requirements[:25]:  # Only show first 25 requirements
+        for requirement in requirements:
             requirement_info = format_requirement(requirement)
             formatted_results.append(requirement_info)
 
@@ -56,7 +56,7 @@ def register_tools(mcp) -> None:
         - Access the full description and selected fields of requirements
 
         Args:
-            product_id: The numeric ID of the product. If the ID is PG:45, just use 45. 
+            product_id: The numeric ID of the product. If the ID is PR:45, just use 45. 
         
         Returns:
             Formatted string containing comprehensive information for the
