+++ src/mcp_server_spira/features/programartifacts/tools/capabilities.py
@@ -4,29 +4,8 @@ Provides operations for working with the Spira program capabilities
 This module provides MCP tools for retrieving program capabilities
 """
 
+from mcp_server_spira.features.formatting import format_capability
 from mcp_server_spira.features.common import get_spira_client
-
-def format_capability(capability) -> str:
-    """
-    Format a program capability into a readable string.
-    
-    Args:
-        capability: The capability data from the API
-        
-    Returns:
-        Formatted string containing the capability information
-    """
-    capability_info = f"""
-## Capability [CP:{capability['CapabilityId']}] - {capability['Name']}
-{'' if capability['Description'] is None else capability['Description']}
-- **Status:** {capability['StatusName']}
-- **Type:** {capability['TypeName']}
-- **Priority:** {capability['PriorityName']}
-- **% Complete:** {capability['PercentComplete']}%
-- **Milestone:** {capability['MilestoneName']}
-- **# Requirements:** {capability['RequirementCount']}
-"""
-    return capability_info
 
 def _get_capabilities_impl(spira_client, program_id: int) -> str:
     """
